paperclipai/pyproject.toml
[project]
name = "paperclipai"
version = "0.1.0"
description = "AI decision service — deterministic agent control plane"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "pydantic[email]>=2.0.0",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0.0",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "sse-starlette>=2.0.0",
    "structlog>=24.0.0",
    "uvicorn[standard]>=0.30.0",
    "honcho>=2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "datamodel-code-generator>=0.26.0",
    "jsonschema[format-nongpl]>=4.23.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "testcontainers[postgres]>=4.8.0",
]
